parse uniqueId for shortcuts, editor improvements

+++ Runtime/Shortcuts/ResolumeOscShortcut.cs
@@ -3,7 +3,7 @@ namespace UnityResolume
     [System.Serializable]
     public class ResolumeOscShortcut
     {
-        public int UniqueId;
+        public long UniqueId;
         public ShortcutPath Input;
         public ShortcutPath Output;
         // not all Shortcuts have subtargets
